refactor(poll): modularize feature, remove demo code, enforce creator-only controls
