+++ firestore.rules
@@ -126,27 +126,39 @@ service cloud.firestore {
         && request.resource.data.votes.size() == request.resource.data.options.size()
         && request.resource.data.status in ['open', 'closed'];
 
-      // Updates:
-      // 1) Anyone may update ONLY the `votes` (and optional `updatedAt`) to support anonymous voting.
-      // 2) Only the creator/host may update `status` and `showLiveResults` (and optional `updatedAt`).
-      // In all cases, question/options/joinCode cannot be modified.
+      // Updates
       allow update: if (
-          // Anonymous voting path
-          request.resource.data.diff(resource.data).changedKeys().hasOnly(['votes', 'updatedAt'])
-          && request.resource.data.votes is list
-          && resource.data.options is list
-          && request.resource.data.votes.size() == resource.data.options.size()
-        )
-        || (
-          // Host controls path
+          // Creator-only controls
           request.auth != null
           && resource.data.createdBy == request.auth.uid
           && request.resource.data.diff(resource.data).changedKeys().hasOnly(['status', 'showLiveResults', 'updatedAt'])
           && request.resource.data.status in ['open', 'closed']
+        ) || (
+          // Authenticated vote write: must also create a vote doc in the same request
+          request.auth != null
+          && request.resource.data.diff(resource.data).changedKeys().hasOnly(['votes', 'updatedAt'])
+          && request.resource.data.votes is list
+          && request.resource.data.votes.size() == resource.data.options.size()
+          && resource.data.status == 'open'
+          && !exists(/databases/$(database)/documents/polls/$(pollId)/votes/$(request.auth.uid))
+          && existsAfter(/databases/$(database)/documents/polls/$(pollId)/votes/$(request.auth.uid))
         );
 
       // No deletes (keeps poll history stable; adjust later if needed)
       allow delete: if false;
+
+      // Votes subcollection: exactly one document per userId
+      match /votes/{userId} {
+        allow read: if request.auth != null && request.auth.uid == userId;
+        allow create: if request.auth != null
+          && request.auth.uid == userId
+          && get(/databases/$(database)/documents/polls/$(pollId)).data.status == 'open'
+          && request.resource.data.optionIndex is int
+          && request.resource.data.optionIndex >= 0
+          && request.resource.data.optionIndex < get(/databases/$(database)/documents/polls/$(pollId)).data.options.size();
+        // Prevent rewrites (one vote per uid)
+        allow update, delete: if false;
+      }
     }
 
     // Block everything else by default
